+++ Firmware/fibre/python/fibre/__init__.py
@@ -1,12 +1,5 @@
 
 from fibre.discovery import find_any, find_all
-from fibre.udp_transport import open_udp
-from fibre.tcp_transport import open_tcp
-try:
-    from fibre.usbbulk_transport import open_usb
-except ModuleNotFoundError:
-    pass
-try:
-    from fibre.serial_transport import open_serial
-except ModuleNotFoundError:
-    pass
+from fibre.utils import Event, Logger
+from fibre.protocol import ChannelBrokenException, ChannelDamagedException
+from fibre.shell import launch_shell
