+++ Firmware/fibre/python/fibre/protocol.py
@@ -2,7 +2,21 @@
 
 import time
 import struct
-from abc import ABC, abstractmethod
+import sys
+import threading
+import traceback
+#import fibre.utils
+from fibre.utils import Event, wait_any
+
+import abc
+if sys.version_info >= (3, 4):
+    ABC = abc.ABC
+else:
+    ABC = abc.ABCMeta('ABC', (), {})
+
+if sys.version_info < (3, 3):
+    from monotonic import monotonic
+    time.monotonic = monotonic
 
 SYNC_BYTE = 0xAA
 CRC8_INIT = 0x42
@@ -30,6 +44,8 @@ def calc_crc(remainder, value, polynomial, bitwidth):
 def calc_crc8(remainder, value):
     if isinstance(value, bytearray) or isinstance(value, bytes) or isinstance(value, list):
         for byte in value:
+            if not isinstance(byte,int):
+                byte = ord(byte)
             remainder = calc_crc(remainder, byte, CRC8_DEFAULT, 8)
     else:
         remainder = calc_crc(remainder, byte, CRC8_DEFAULT, 8)
@@ -38,6 +54,8 @@ def calc_crc8(remainder, value):
 def calc_crc16(remainder, value):
     if isinstance(value, bytearray) or isinstance(value, bytes) or isinstance(value, list):
         for byte in value:
+            if not isinstance(byte, int):
+                byte = ord(byte)
             remainder = calc_crc(remainder, byte, CRC16_DEFAULT, 16)
     else:
         remainder = calc_crc(remainder, value, CRC16_DEFAULT, 16)
@@ -48,43 +66,49 @@ def calc_crc16(remainder, value):
 #print(hex(calc_crc16(0xfeef, [1, 2, 3, 4, 5, 0x10, 0x13, 0x37])))
 
 
-class TimeoutException(Exception):
+class DeviceInitException(Exception):
+    pass
+
+class ChannelDamagedException(Exception):
+    """
+    Raised when the channel is temporarily broken and a
+    resend of the message might be successful
+    """
     pass
 
 class ChannelBrokenException(Exception):
-    pass
-
-class DeviceInitException(Exception):
+    """
+    Raised when the channel is permanently broken
+    """
     pass
 
 
 class StreamSource(ABC):
-    @abstractmethod
+    @abc.abstractmethod
     def get_bytes(self, n_bytes, deadline):
         pass
 
 class StreamSink(ABC):
-    @abstractmethod
+    @abc.abstractmethod
     def process_bytes(self, bytes):
         pass
 
 class PacketSource(ABC):
-    @abstractmethod
+    @abc.abstractmethod
     def get_packet(self, deadline):
         pass
 
 class PacketSink(ABC):
-    @abstractmethod
+    @abc.abstractmethod
     def process_packet(self, packet):
         pass
 
 
 class StreamToPacketSegmenter(StreamSink):
-    _header = []
-    _packet = []
-    _packet_length = 0
-
     def __init__(self, output):
+        self._header = []
+        self._packet = []
+        self._packet_length = 0
         self._output = output
 
     def process_bytes(self, bytes):
@@ -140,9 +164,8 @@ class StreamBasedPacketSink(PacketSink):
         self._output.process_bytes(struct.pack('>H', crc16))
 
 class PacketFromStreamConverter(PacketSource):
-    def __init__(self, input, trash):
+    def __init__(self, input):
         self._input = input
-        self._trash = trash
     
     def get_packet(self, deadline):
         """
@@ -157,19 +180,16 @@ class PacketFromStreamConverter(PacketSource):
             header = header + self._input.get_bytes_or_fail(1, deadline)
             if (header[0] != SYNC_BYTE):
                 #print("sync byte mismatch")
-                self._trash(header)
                 continue
 
             header = header + self._input.get_bytes_or_fail(1, deadline)
             if (header[1] & 0x80):
                 #print("packet too large")
-                self._trash(header)
                 continue # TODO: support packets larger than 128 bytes
 
             header = header + self._input.get_bytes_or_fail(1, deadline)
             if calc_crc8(CRC8_INIT, header) != 0:
                 #print("crc8 mismatch")
-                self._trash(header)
                 continue
 
             packet_length = header[1] + 2
@@ -177,22 +197,16 @@ class PacketFromStreamConverter(PacketSource):
             packet = self._input.get_bytes_or_fail(packet_length, deadline)
             if calc_crc16(CRC16_INIT, packet) != 0:
                 #print("crc16 mismatch")
-                self._trash(header)
-                self._trash(packet)
                 continue
             return packet[:-2]
 
 
 class Channel(PacketSink):
-    _outbound_seq_no = 0
-    _interface_definition_crc = 0
-    _expected_acks = {}
-
     # Choose these parameters to be sensible for a specific transport layer
     _resend_timeout = 5.0     # [s]
     _send_attempts = 5
 
-    def __init__(self, name, input, output, trash):
+    def __init__(self, name, input, output, cancellation_token, logger):
         """
         Params:
         input: A PacketSource where this channel will source packets from on
@@ -203,7 +217,46 @@ class Channel(PacketSink):
         self._name = name
         self._input = input
         self._output = output
-        self._trash = trash
+        self._logger = logger
+        self._outbound_seq_no = 0
+        self._interface_definition_crc = 0
+        self._expected_acks = {}
+        self._responses = {}
+        self._my_lock = threading.Lock()
+        self._channel_broken = Event(cancellation_token)
+        self.start_receiver_thread(Event(self._channel_broken))
+
+    def start_receiver_thread(self, cancellation_token):
+        """
+        Starts the receiver thread that processes incoming messages.
+        The thread quits as soon as the channel enters a broken state.
+        """
+        def receiver_thread():
+            error_ctr = 0
+            try:
+                while (not cancellation_token.is_set() and not self._channel_broken.is_set()
+                        and error_ctr < 10):
+                    # Set an arbitrary deadline because the get_packet function
+                    # currently doesn't support a cancellation_token
+                    deadline = time.monotonic() + 1.0
+                    try:
+                        response = self._input.get_packet(deadline)
+                    except TimeoutError:
+                        continue # try again
+                    except ChannelDamagedException:
+                        error_ctr += 1
+                        continue # try again
+                    if (error_ctr > 0):
+                        error_ctr -= 1
+                    # Process response
+                    # This should not throw an exception, otherwise the channel breaks
+                    self.process_packet(response)
+                #print("receiver thread is exiting")
+            except Exception:
+                self._logger.debug("receiver thread is exiting: " + traceback.format_exc())
+            finally:
+                self._channel_broken.set()
+        threading.Thread(target=receiver_thread).start()
 
     def remote_endpoint_operation(self, endpoint_id, input, expect_ack, output_length):
         if input is None:
@@ -214,9 +267,13 @@ class Channel(PacketSink):
         if (expect_ack):
             endpoint_id |= 0x8000
 
-        self._outbound_seq_no = ((self._outbound_seq_no + 1) & 0x7fff)
-        self._outbound_seq_no |= 0x80 # FIXME: we hardwire one bit of the seq-no to 1 to avoid conflicts with the legacy protocol
-        seq_no = self._outbound_seq_no
+        self._my_lock.acquire()
+        try:
+            self._outbound_seq_no = ((self._outbound_seq_no + 1) & 0x7fff)
+            seq_no = self._outbound_seq_no
+        finally:
+            self._my_lock.release()
+        seq_no |= 0x80 # FIXME: we hardwire one bit of the seq-no to 1 to avoid conflicts with the ascii protocol
         packet = struct.pack('<HHH', seq_no, endpoint_id, output_length)
         packet = packet + input
 
@@ -229,25 +286,32 @@ class Channel(PacketSink):
         packet = packet + struct.pack('<H', trailer)
 
         if (expect_ack):
-            self._expected_acks[seq_no] = None
-            attempt = 0
-            while (attempt < self._send_attempts):
-                self._output.process_packet(packet)
-                deadline = time.monotonic() + self._resend_timeout
-                # Read and process packets until we get an ack or need to resend
-                # TODO: support I/O driven reception (wait on semaphore)
-                while True:
+            ack_event = Event()
+            self._expected_acks[seq_no] = ack_event
+            try:
+                attempt = 0
+                while (attempt < self._send_attempts):
+                    self._my_lock.acquire()
                     try:
-                        response = self._input.get_packet(deadline)
-                    except TimeoutException:
-                        break # resend
-                    # process response, which is hopefully our ACK
-                    self.process_packet(response)
-                    if not self._expected_acks[seq_no] is None:
-                        return self._expected_acks.pop(seq_no, None)
-                # TODO: record channel statistics
-                attempt += 1
-            raise ChannelBrokenException()
+                        self._output.process_packet(packet)
+                    except ChannelDamagedException:
+                        attempt += 1
+                        continue # resend
+                    finally:
+                        self._my_lock.release()
+                    # Wait for ACK until the resend timeout is exceeded
+                    try:
+                        if wait_any(self._resend_timeout, ack_event, self._channel_broken) != 0:
+                            raise ChannelBrokenException()
+                    except TimeoutError:
+                        attempt += 1
+                        continue # resend
+                    return self._responses.pop(seq_no)
+                    # TODO: record channel statistics
+                raise ChannelBrokenException() # Too many resend attempts
+            finally:
+                self._expected_acks.pop(seq_no)
+                self._responses.pop(seq_no, None)
         else:
             # fire and forget
             self._output.process_packet(packet)
@@ -260,7 +324,7 @@ class Channel(PacketSink):
         # TODO: handle device that could (maliciously) send infinite stream
         buffer = bytes()
         while True:
-            chunk_length = 64
+            chunk_length = 512
             chunk = self.remote_endpoint_operation(endpoint_id, struct.pack("<I", len(buffer)), True, chunk_length)
             if (len(chunk) == 0):
                 break
@@ -277,13 +341,16 @@ class Channel(PacketSink):
 
         if (seq_no & 0x8000):
             seq_no &= 0x7fff
-            self._expected_acks[seq_no] = packet[2:]
+            ack_signal = self._expected_acks.get(seq_no, None)
+            if (ack_signal):
+                self._responses[seq_no] = packet[2:]
+                ack_signal.set()
+                #print("received ack for packet " + str(seq_no))
+            else:
+                print("received unexpected ACK: " + str(seq_no))
 
         else:
-            #if (calc_crc16(crc16, struct.pack('<HBB', PROTOCOL_VERSION, packet[-2], packet[-1]))):
-            #    raise Exception("CRC16 mismatch")
-            #print("endpoint requested")
-            # FIXME: we use non-ack packets to detect printf output, this is really hacky
-            # In the future there should be a dedicated stdout endpoint.
-            self._trash(packet)
+            #if (calc_crc16(CRC16_INIT, struct.pack('<HBB', PROTOCOL_VERSION, packet[-2], packet[-1]))):
+            #     raise Exception("CRC16 mismatch")
+            print("endpoint requested")
             # TODO: handle local endpoint operation
