+++ Firmware/fibre/tools/fibre-shell
@@ -1,0 +1,57 @@
+#!/usr/bin/env python3
+"""
+Connect to a Fibre-enabled device to play with in the IPython interactive shell.
+"""
+import argparse
+import sys
+import os
+
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.realpath(__file__))) + "/python")
+import fibre
+from fibre import Logger, Event
+
+# Parse arguments
+parser = argparse.ArgumentParser(description='Connect to a fibre-enabled device to play with it in the IPython interactive shell.')
+parser.add_argument("-p", "--path", metavar="PATH", action="store",
+                    help="The path(s) where ODrive(s) should be discovered.\n"
+                    "By default the script will connect to any ODrive on USB.\n\n"
+                    "To select a specific USB device:\n"
+                    "  --path usb:BUS:DEVICE\n"
+                    "usbwhere BUS and DEVICE are the bus and device numbers as shown in `lsusb`.\n\n"
+                    "To select a specific serial port:\n"
+                    "  --path serial:PATH\n"
+                    "where PATH is the path of the serial port. For example \"/dev/ttyUSB0\".\n"
+                    "You can use `ls /dev/tty*` to find the correct port.\n\n"
+                    "You can combine USB and serial specs by separating them with a comma (no space!)\n"
+                    "Example:\n"
+                    "  --path usb,serial:/dev/ttyUSB0\n"
+                    "means \"discover any USB device or a serial device on /dev/ttyUSB0\"")
+parser.add_argument("-s", "--serial-number", action="store",
+                    help="The 12-digit serial number of the device. "
+                         "This is a string consisting of 12 upper case hexadecimal "
+                         "digits as displayed in lsusb. \n"
+                         "    example: 385F324D3037\n"
+                         "You can list all devices connected to USB by running\n"
+                         "(lsusb -d 1209:0d32 -v; lsusb -d 0483:df11 -v) | grep iSerial\n"
+                         "If omitted, any device is accepted.")
+parser.add_argument("--no-ipython", action="store_true",
+                    help="Use the regular Python shell "
+                         "instead of the IPython shell, "
+                         "even if IPython is installed.")
+parser.add_argument("-v", "--verbose", action="store_true",
+                    help="print debug information")
+
+parser.set_defaults(path="usb,tcp:localhost:9910")
+args = parser.parse_args()
+
+logger = Logger(verbose=args.verbose)
+app_shutdown_token = Event()
+
+def print_banner():
+    pass
+
+def print_help(args, have_devices):
+    pass
+
+import fibre
+fibre.launch_shell(args, print_banner, print_help, logger, app_shutdown_token)
