+++ tools/odrive/__init__.py
@@ -3,3 +3,8 @@
 from .version import get_version_str
 __version__ = get_version_str()
 del get_version_str
+
+
+import fibre
+find_any = fibre.find_any
+find_all = fibre.find_all
